requirements.txt
# Core utilities
python-dotenv==1.0.0
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Web frameworks
streamlit==1.29.0
flask==3.0.0

# AI/ML
openai==1.6.1
anthropic==0.8.1

# Bots
python-telegram-bot==20.7
discord.py==2.3.2

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0

# Development tools
black==23.12.1
flake8==7.0.0
pytest==7.4.3
pre-commit==3.6.0
